Simple version of category rankings
This essentially fixes the issue but I would much rather have a nicer
presentation and a more usable interface.

NOTE: We need to stop having repeates in admin index and report
index. Maybe consolidate the lists somehow?

The whole product display table is also not very thought out since it
repeates all of the code every single time. Lets create a single
template for that stuff.