More enforceable limits
This patch gets around the transaction on by issue by making a limit
a quantity and a start date. When we have to find out how many we have
already sold we then just tally up the number of sales if this product
since the start date. The query to do so is slightly complicated, but it
works as it stands.

This is more robust and less error prone than the method already merged.

It does have the potential to launch a lot of queries, we will have to
monitor this when we deploy!

Relates to #73 and #23