Faster limit query
The other one was just terrible. I had successfully limited myself to
once query, the price was that it had to go over every single sale in
the database, instead of just the ones newer than the start_date. It was
dog slow, and only going to get slow.

The new one here does between 1-3 queries in the db, but should be much
quicker. I've tested it on the complete dataset of the running system,
an it seem performant enough on my laptop.