And an index to sale product_id,timestamp
This should resolve the issue with slow queries. Admittedly I have no
idea how to come to this conclusion based on analysis, but i tried it
and it's pretty fast.

We could technically do the out_of_order query in python, which is still
pretty fast. Testing shows this to be faster though!