Improve API (API 1.1) (#421)
* Streamline GetMemberID with other Member-API endpoints

The other endpoints use 'Invalid ID' to signify either missing or
invalid. This endpoint used to use 'Invalid Username' to signify that the
member does not exist, while the other's would've used 'Member not
found' for that.

* Consistent member/user terminology

There's an inconsistency in whether to call a member a user. Generally,
users are the ones with admin access, while members are the customers.

* Consistent view-method naming

Even though it's a bool, I believe it's more appropriate to keep the
naming consistent with non-bool return types. Also, it still returns an
object.

* Make get ID consistent as well

I won't rename the actual endpoint because of backwards compatibility.

* Rename Get Payment QR-endpoint

* Refactor Get Payment QR

* Rename named_products method

* Change return text

* Remove use of annotations from future

* Object instead of array in Active Products-endpoint

More meaningful response in Active Products-endpoint, now returns {"1":
{"name": "Beer", "price": 600}}, instead of {"1": ["Beer", 600]}

* TODO Check whether room exists

* Rename URL name

* Strongly typed return-object for category mapping

* Make QR Form and helper function generic

* Rename 'member' -> 'username'-parameter

* Change minimum value for QR-code form to 0

* Fix incorrect 400 response body's

* Make api/sale-endpoint responses more specific

* Remove unused import

* Rename view-functions

* Change naming scheme